science: add ScienceMap component
This new component will also eventually live at `/science/map` with a
smaller version of the map living at `/science`.

This map, unlike the FluMap at `/current`, is intended for a public
health or epidemiological audience rather than the general public.

Much of the code here is copied and pasted from the FluMap component,
and with handlers for keyframes, it is perhaps over-engineered for this
particular use-case.

The styling for the heatmap is copied from Trevor's
[heatmap](seattleflu/heatmap).

The data are currently sourced from a static file export of [this
Metabase
query](https://backoffice.seattleflu.org/metabase/question/281).