Unify our React "subapps" into a single build that produces multiple bundles
This greatly simplifies development at the potential cost of not having
isolated dependencies because one package.json is shared by everything.
I think this is worth it.  While having entirely independent apps seemed
like a good idea initially, in practice it worked poorly and involved a
lot of extra complexity for ~zero gain.

I'm pretty happy with the net of this diff stat:

    4884 insertions(+), 31415 deletions(-)