Current: Map shows scaled IDM model results over Seattle neighborhoods
Lots of changes, this is not a well-factored commit, but it does only
touch the FluMap component.

Particularly, some complex handling of multiple data layers was
discarded since we only need to show a single data layer (neighborhoods
now, later combined neighborhoods + PUMAs on a single layer).  Code is
still mostly well-factored so adding additional data layers won't be too
hard.  The new <Layer> and <Source> components for <MapboxGL> made this
a nice win, particularly since we have to thread date updates into the
extrusion styling.