Current: Rename data source-related names to be consistent
I was looking over this again and noticed that while "neighborhoods()"
made sense in an earlier iteration of the code, it makes less sense now
as it is the only data layer.  If we change the displayed geometry,
we'll change this function in-place rather than make a new function.

Aligns names on "dataSource" so data flow is easier to trace when
reading.  The previous "dataLayer" state was a bit misleading since it
is used directly by <Source> not <Layer>.