Current: Improve drag and touch behavior
The previous behavior would often "capture" an intent to scroll and instead zoom the map. I've sought to replicate the behavior of darksky.net where scroll does not zoom.

There is still an issue with mouse drag vs single finger pan. react-map-gl doesn't allow these to be separated. I've opted to keep this interaction. The map can be captured on mobile. Because of this I've slightly shrunk the vertical expanse to make it not take up the entire mobile display. 

When nav icons are restored, we should provide zoom in / zoom out controls.