results: Improve error logging
Currently error logging occurs with the word `ERROR` and a console print
out of the error. The browser console traces these errors back to
`results.js`, but this could be the file under `routes/` or `services/`.
Disambiguate the file name and identify the function that errored out by
adding a better description before the error print out.