Add Pathogens page
Embeds a chart and caption from an Observable notebook.¹  Other page
data comes from Contentful, as usual.

Embedding from Observable was the quickest way to get this page
together.  It's not my normal approach, but I think it'll be the most
maintenance-free approach.  We _could_ extract the code in the notebook
and use it directly in the website here, but we'd lose (or have to
re-implement) things like auto-resizing, dep management, and other
things that Observable handles for us.  It's just more overhead, for
benefit I'm not sure we need.

The notebook is currently associated with my personal Observable
account, but we can remedy this in the (near) future by forking the
notebook elsewhere (a shared "personal" account or an Observable team
account) and updating the embed URL here to point to the fork.  I didn't
want to block getting the page up on resolving the notebook ownership.

I've created and published the "Pathogens" page in the "prod"
environment in Contentful since it won't affect the live site and makes
previewing in dev easy.  Once this commit is deployed, we'll want to add
the Pathogens page to the menu as well, which is managed in Contentful.

Except for the embedding code, the rest of the changes here are the
copypasta that's typical for this repo.

¹ https://observablehq.com/@tsibley/pathogen-prevalence