pathogens: Add specimens chart to page
Embeds from a new forked copy of the original notebook.  Some
interaction behaviour for the prevalence chart is updated as well.

A dedicated element is used as the "ruler" to measure container width
for both charts, since I changed the embed hooks slightly which made it
awkward to keep the "chart" reference around.